fixed kotlin error in button callback
