+++ library/build.gradle
@@ -17,6 +17,7 @@ apply from: '../dependencies.gradle'
 group='com.github.AdityaAnand1'
 
 android {
+    kapt { generateStubs = true }
     compileSdkVersion versions.compileSdk
     buildToolsVersion versions.buildTools
 
@@ -47,6 +48,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
+    kapt "com.android.databinding:compiler:${versions.gradlePlugin}"
     implementation "com.android.support:appcompat-v7:${versions.supportLib}"
     implementation "com.android.support:design:${versions.supportLib}"
     implementation 'com.afollestad.material-dialogs:core:0.9.6.0'
