+++ library/proguard-rules.pro
@@ -26,7 +26,7 @@
 -dontwarn java.lang.invoke**
 -dontwarn android.databinding.**
 -keep class android.databinding.** { *; }
--keep class in.adityaanand.morphdialog.databinding.** {
+-keep class com.adityaanand.morphdialog.databinding.** {
     <fields>;
     <methods>;
 }
