+++ library/src/main/AndroidManifest.xml
@@ -1,2 +1,2 @@
-<manifest package="in.adityaanand.morphdialog"
+<manifest package="com.adityaanand.morphdialog"
           xmlns:android="http://schemas.android.com/apk/res/android"/>
