+++ library/src/main/java/com/adityaanand/morphdialog/interfaces/MorphSingleButtonCallback.kt
@@ -1,0 +1,9 @@
+package com.adityaanand.morphdialog.interfaces
+
+import com.adityaanand.morphdialog.MorphDialog
+import com.adityaanand.morphdialog.utils.MorphDialogAction
+
+interface MorphSingleButtonCallback {
+
+    fun onClick(dialog: MorphDialog, which: MorphDialogAction)
+}
