+++ library/src/main/java/com/adityaanand/morphdialog/morphutil/MorphDialogToFab.java
@@ -14,47 +14,44 @@
  * limitations under the License.
  */
 
-package in.adityaanand.morphdialog.morphutil;
+package com.adityaanand.morphdialog.morphutil;
 
 import android.animation.Animator;
 import android.animation.AnimatorSet;
 import android.animation.ObjectAnimator;
 import android.annotation.TargetApi;
 import android.content.Context;
-import android.content.res.TypedArray;
 import android.os.Build;
 import android.transition.ChangeBounds;
 import android.transition.TransitionValues;
 import android.util.AttributeSet;
-import android.util.TypedValue;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.AnimationUtils;
 
-import in.adityaanand.morphdialog.R;
+import com.adityaanand.morphdialog.R;
 
 
 /**
- * A transition that morphs a circle into a rectangle, changing it's background color.
+ * A transition that morphs a rectangle into a circle, changing it's background color.
  */
 @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-public class MorphFabToDialog extends ChangeBounds {
+public class MorphDialogToFab extends ChangeBounds {
 
-    private static final String PROPERTY_COLOR = "circleMorph:color";
-    private static final String PROPERTY_CORNER_RADIUS = "circleMorph:cornerRadius";
+    private static final String PROPERTY_COLOR = "rectMorph:color";
+    private static final String PROPERTY_CORNER_RADIUS = "rectMorph:cornerRadius";
     private static final String[] TRANSITION_PROPERTIES = {
             PROPERTY_COLOR,
             PROPERTY_CORNER_RADIUS
     };
     int backgroundColor;
 
-    public MorphFabToDialog(int backgroundColor) {
+    public MorphDialogToFab(int backgroundColor) {
         super();
         this.backgroundColor = backgroundColor;
     }
 
-
-    public MorphFabToDialog(Context context, AttributeSet attrs) {
+    public MorphDialogToFab(Context context, AttributeSet attrs) {
         super(context, attrs);
     }
 
@@ -70,19 +67,9 @@ public class MorphFabToDialog extends ChangeBounds {
         if (view.getWidth() <= 0 || view.getHeight() <= 0) {
             return;
         }
-        transitionValues.values.put(PROPERTY_COLOR, fetchAccentColor(view.getContext()));
-        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getHeight() / 2);//view.getHeight() / 2
-    }
-
-    public static int fetchAccentColor(Context context) {
-        TypedValue typedValue = new TypedValue();
-
-        TypedArray a = context.obtainStyledAttributes(typedValue.data, new int[]{R.attr.colorAccent});
-        int color = a.getColor(0, 0);
-
-        a.recycle();
-
-        return color;
+        transitionValues.values.put(PROPERTY_COLOR, backgroundColor);
+        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getResources()
+                .getDimensionPixelSize(R.dimen.default_dialog_corners));
     }
 
     @Override
@@ -92,12 +79,12 @@ public class MorphFabToDialog extends ChangeBounds {
         if (view.getWidth() <= 0 || view.getHeight() <= 0) {
             return;
         }
-        transitionValues.values.put(PROPERTY_COLOR, backgroundColor);
-        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getResources().getDimensionPixelSize(R.dimen.default_dialog_corners));
+        transitionValues.values.put(PROPERTY_COLOR, MorphFabToDialog.fetchAccentColor(view.getContext()));
+        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getHeight() / 2);//view.getHeight() / 2
     }
 
     @Override
-    public Animator createAnimator(final ViewGroup sceneRoot, TransitionValues startValues, final TransitionValues endValues) {
+    public Animator createAnimator(final ViewGroup sceneRoot, TransitionValues startValues, TransitionValues endValues) {
         Animator changeBounds = super.createAnimator(sceneRoot, startValues, endValues);
         if (startValues == null || endValues == null || changeBounds == null) {
             return null;
@@ -118,25 +105,21 @@ public class MorphFabToDialog extends ChangeBounds {
         Animator color = ObjectAnimator.ofArgb(background, MorphDrawable.COLOR, endColor);
         Animator corners = ObjectAnimator.ofFloat(background, MorphDrawable.CORNER_RADIUS, endCornerRadius);
 
-        // ease in the dialog's child views (slide up & fade in)
+        // hide child views (offset down & fade out)
         if (endValues.view instanceof ViewGroup) {
             ViewGroup vg = (ViewGroup) endValues.view;
-            float offset = vg.getHeight() / 3;
             for (int i = 0; i < vg.getChildCount(); i++) {
                 View v = vg.getChildAt(i);
-                v.setTranslationY(offset);
-                v.setAlpha(0f);
-                v.animate().alpha(1f).translationY(0f).setDuration(150).setStartDelay(150)
-                        .setInterpolator(AnimationUtils.loadInterpolator(vg.getContext(), android.R.interpolator.fast_out_slow_in))
+                v.animate().alpha(0f).translationY(v.getHeight() / 3).setStartDelay(0L).setDuration(50L)
+                        .setInterpolator(AnimationUtils.loadInterpolator(vg.getContext(), android.R.interpolator.fast_out_linear_in))
                         .start();
-                offset *= 1.8f;
             }
         }
 
         AnimatorSet transition = new AnimatorSet();
         transition.playTogether(changeBounds, corners, color);
-        transition.setDuration(300);
         transition.setInterpolator(AnimationUtils.loadInterpolator(sceneRoot.getContext(), android.R.interpolator.fast_out_slow_in));
+        transition.setDuration(300);
         return transition;
     }
 }
