+++ library/src/main/java/com/adityaanand/morphdialog/utils/MorphDialogAction.kt
@@ -1,4 +1,4 @@
-package `in`.adityaanand.morphdialog.utils
+package com.adityaanand.morphdialog.utils
 
 enum class MorphDialogAction {
     POSITIVE,
