--- library/src/main/java/in/adityaanand/morphdialog/interfaces/MorphSingleButtonCallback.kt
@@ -1,9 +1,0 @@
-package `in`.adityaanand.morphdialog.interfaces
-
-import `in`.adityaanand.morphdialog.MorphDialog
-import `in`.adityaanand.morphdialog.utils.MorphDialogAction
-
-interface MorphSingleButtonCallback {
-
-    fun onClick(dialog: MorphDialog, which: MorphDialogAction)
-}
