+++ standalone/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="in.adityaanand.morphdialogstandalone"
+<manifest package="com.adityaanand.morphdialogstandalone"
           xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
@@ -9,7 +9,7 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppBaseTheme">
-        <activity android:name="in.adityaanand.morphdialogstandalone.MainActivity">
+        <activity android:name="com.adityaanand.morphdialogstandalone.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
 
@@ -17,7 +17,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name="in.adityaanand.morphdialogstandalone.DialogActivity"
+            android:name="com.adityaanand.morphdialogstandalone.DialogActivity"
             android:theme="@style/MorphDialog.Light">
         </activity>
     </application>
