+++ standalone/src/main/java/com/adityaanand/morphdialogstandalone/MainActivity.java
@@ -1,11 +1,11 @@
-package in.adityaanand.morphdialogstandalone;
+package com.adityaanand.morphdialogstandalone;
 
 import android.databinding.DataBindingUtil;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import in.adityaanand.morphdialogstandalone.databinding.ActivityMainBinding;
+import com.adityaanand.morphdialogstandalone.databinding.ActivityMainBinding;
 
 public class MainActivity extends AppCompatActivity {
