+++ standalone/src/main/java/com/adityaanand/morphdialogstandalone/util/MorphDialogToFab.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package in.adityaanand.morphdialogstandalone.util;
+package com.adityaanand.morphdialogstandalone.util;
 
 import android.animation.Animator;
 import android.animation.AnimatorSet;
@@ -30,30 +30,29 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.AnimationUtils;
 
-import in.adityaanand.morphdialogstandalone.R;
+import com.adityaanand.morphdialogstandalone.R;
 
 
 /**
- * A transition that morphs a circle into a rectangle, changing it's background color.
+ * A transition that morphs a rectangle into a circle, changing it's background color.
  */
 @TargetApi(Build.VERSION_CODES.LOLLIPOP)
-public class MorphFabToDialog extends ChangeBounds {
+public class MorphDialogToFab extends ChangeBounds {
 
-    private static final String PROPERTY_COLOR = "circleMorph:color";
-    private static final String PROPERTY_CORNER_RADIUS = "circleMorph:cornerRadius";
+    private static final String PROPERTY_COLOR = "rectMorph:color";
+    private static final String PROPERTY_CORNER_RADIUS = "rectMorph:cornerRadius";
     private static final String[] TRANSITION_PROPERTIES = {
             PROPERTY_COLOR,
             PROPERTY_CORNER_RADIUS
     };
     int backgroundColor;
 
-    public MorphFabToDialog(int backgroundColor) {
+    public MorphDialogToFab(int backgroundColor) {
         super();
         this.backgroundColor = backgroundColor;
     }
 
-
-    public MorphFabToDialog(Context context, AttributeSet attrs) {
+    public MorphDialogToFab(Context context, AttributeSet attrs) {
         super(context, attrs);
     }
 
@@ -69,8 +68,9 @@ public class MorphFabToDialog extends ChangeBounds {
         if (view.getWidth() <= 0 || view.getHeight() <= 0) {
             return;
         }
-        transitionValues.values.put(PROPERTY_COLOR, ContextCompat.getColor(view.getContext(), R.color.accent));
-        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getHeight() / 2);//view.getHeight() / 2
+        transitionValues.values.put(PROPERTY_COLOR, backgroundColor);
+        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getResources()
+                .getDimensionPixelSize(R.dimen.dialog_corners));
     }
 
     @Override
@@ -80,12 +80,12 @@ public class MorphFabToDialog extends ChangeBounds {
         if (view.getWidth() <= 0 || view.getHeight() <= 0) {
             return;
         }
-        transitionValues.values.put(PROPERTY_COLOR, backgroundColor);
-        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getResources().getDimensionPixelSize(R.dimen.dialog_corners));
+        transitionValues.values.put(PROPERTY_COLOR, ContextCompat.getColor(view.getContext(), R.color.accent));
+        transitionValues.values.put(PROPERTY_CORNER_RADIUS, view.getHeight() / 2);//view.getHeight() / 2
     }
 
     @Override
-    public Animator createAnimator(final ViewGroup sceneRoot, TransitionValues startValues, final TransitionValues endValues) {
+    public Animator createAnimator(final ViewGroup sceneRoot, TransitionValues startValues, TransitionValues endValues) {
         Animator changeBounds = super.createAnimator(sceneRoot, startValues, endValues);
         if (startValues == null || endValues == null || changeBounds == null) {
             return null;
@@ -106,25 +106,21 @@ public class MorphFabToDialog extends ChangeBounds {
         Animator color = ObjectAnimator.ofArgb(background, MorphDrawable.COLOR, endColor);
         Animator corners = ObjectAnimator.ofFloat(background, MorphDrawable.CORNER_RADIUS, endCornerRadius);
 
-        // ease in the dialog's child views (slide up & fade in)
+        // hide child views (offset down & fade out)
         if (endValues.view instanceof ViewGroup) {
             ViewGroup vg = (ViewGroup) endValues.view;
-            float offset = vg.getHeight() / 3;
             for (int i = 0; i < vg.getChildCount(); i++) {
                 View v = vg.getChildAt(i);
-                v.setTranslationY(offset);
-                v.setAlpha(0f);
-                v.animate().alpha(1f).translationY(0f).setDuration(150).setStartDelay(150)
-                        .setInterpolator(AnimationUtils.loadInterpolator(vg.getContext(), android.R.interpolator.fast_out_slow_in))
+                v.animate().alpha(0f).translationY(v.getHeight() / 3).setStartDelay(0L).setDuration(50L)
+                        .setInterpolator(AnimationUtils.loadInterpolator(vg.getContext(), android.R.interpolator.fast_out_linear_in))
                         .start();
-                offset *= 1.8f;
             }
         }
 
         AnimatorSet transition = new AnimatorSet();
         transition.playTogether(changeBounds, corners, color);
-        transition.setDuration(300);
         transition.setInterpolator(AnimationUtils.loadInterpolator(sceneRoot.getContext(), android.R.interpolator.fast_out_slow_in));
+        transition.setDuration(300);
         return transition;
     }
 }
